Re-added toolbar to gpodder main fragment
and removed unused gpodder search fragment
