+++ app/src/main/res/layout/gpodnet_podcast_list.xml
@@ -2,17 +2,8 @@
 <RelativeLayout
     xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:tools="http://schemas.android.com/tools"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
     android:layout_width="match_parent"
     android:layout_height="match_parent">
-
-    <androidx.appcompat.widget.Toolbar
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:minHeight="?attr/actionBarSize"
-        android:theme="?attr/actionBarTheme"
-        app:navigationIcon="?homeAsUpIndicator"
-        android:id="@+id/toolbar"/>
 
     <GridView
         android:id="@+id/gridView"
@@ -27,7 +18,7 @@
         android:paddingTop="@dimen/list_vertical_padding"
         android:stretchMode="columnWidth"
         android:verticalSpacing="8dp"
-        android:layout_below="@id/toolbar"
+        android:layout_alignParentTop="true"
         tools:listitem="@layout/gpodnet_podcast_listitem" />
 
     <ProgressBar
