--- app/src/main/res/menu/gpodder_podcasts.xml
@@ -1,13 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-
-<menu xmlns:android="http://schemas.android.com/apk/res/android"
-      xmlns:custom="http://schemas.android.com/apk/res-auto">
-
-    <item
-        android:id="@+id/action_search"
-        android:icon="?attr/action_search"
-        custom:showAsAction="collapseActionView|ifRoom"
-        custom:actionViewClass="androidx.appcompat.widget.SearchView"
-        android:title="@string/search_label"/>
-
-</menu>
